Remove habitData.js, revert Habit.svelte, inline ContributionGraph fixes only
